requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
qdrant-client
sentence-transformers
python-multipart
alembic